requirements.txt
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.27
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
faiss-cpu==1.9.0
pypdf==5.1.0
python-dotenv==1.0.1
sentence-transformers==3.3.0
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12